TA#50763 [ADD] hr_employee_declaration : adding new module
