--- .unported/hr_employee_declaration/tests/__init__.py
@@ -1,2 +1,0 @@
-# © 2018 - today Numigi (tm) and all its contributors (https://bit.ly/numigiens)
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
