+++ Dockerfile
@@ -18,7 +18,7 @@ COPY hr_attendance_tracking_visibility /mnt/extra-addons/hr_attendance_tracking_
 COPY hr_contract_single_open /mnt/extra-addons/hr_contract_single_open
 COPY hr_contract_wage_hourly /mnt/extra-addons/hr_contract_wage_hourly
 COPY hr_contract_wage_type /mnt/extra-addons/hr_contract_wage_type
-#COPY hr_employee_declaration /mnt/extra-addons/hr_employee_declaration
+COPY hr_employee_declaration /mnt/extra-addons/hr_employee_declaration
 COPY hr_employee_private_wizard /mnt/extra-addons/hr_employee_private_wizard
 COPY hr_employee_type /mnt/extra-addons/hr_employee_type
 COPY hr_timesheet_enhanced /mnt/extra-addons/hr_timesheet_enhanced
