+++ hr_employee_declaration/__init__.py
@@ -1,0 +1,4 @@
+# © 2023 Numigi (tm) and all its contributors (https://bit.ly/numigiens)
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+from . import  models
