+++ hr_employee_declaration/__manifest__.py
@@ -1,9 +1,9 @@
-# © 2018 - today Numigi (tm) and all its contributors (https://bit.ly/numigiens)
+# © 2023 Numigi (tm) and all its contributors (https://bit.ly/numigiens)
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
 {
     'name': 'Employee Declarations',
-    'version': '1.0.0',
+    'version': '14.0.1.0.0',
     'author': 'Numigi',
     'maintainer': 'Numigi',
     'website': 'https://bit.ly/numigi-com',
@@ -18,6 +18,8 @@
     'data': [
         'data/survey_type.xml',
         'data/ir_cron.xml',
+        'data/queue_job_function_data.xml',
+        #'demo/survey.xml',
         'views/hr_employee_with_declaration_smart_button.xml',
         'views/hr_employee_with_declaration_settings.xml',
         'views/survey_type_with_is_employee_declaration.xml',
