+++ hr_employee_declaration/data/queue_job_function_data.xml
@@ -1,0 +1,14 @@
+<odoo noupdate="0">
+    <record model="queue.job.channel" id="channel_hr_employee">
+        <field name="name">hr_employee</field>
+        <field name="parent_id" ref="queue_job.channel_root" />
+    </record>
+
+    <record id="job_function_send_declaration_survey_by_email" model="queue.job.function">
+        <field name="model_id" ref="hr.model_hr_employee" />
+        <field name="method">send_declaration_survey_by_email</field>
+        <field name="channel_id" ref="channel_hr_employee" />
+
+    </record>
+</odoo>
+
