+++ hr_employee_declaration/data/survey_type.xml
