+++ hr_employee_declaration/i18n/fr.po
