+++ hr_employee_declaration/models/__init__.py
@@ -1,0 +1,6 @@
+# © 2023 Numigi (tm) and all its contributors (https://bit.ly/numigiens)
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+from . import employee_with_declarations
+from . import send_declaration_survey_by_mail
+from . import send_declaration_survey_cron
