+++ hr_employee_declaration/security/ir_rule.xml
