+++ hr_employee_declaration/static/description/employee_declaration_tab.png
