+++ hr_employee_declaration/static/description/employee_email_template_field.png
