+++ hr_employee_declaration/static/description/employee_periodic_declaration_field.png
