+++ hr_employee_declaration/static/description/employee_send_to_field.png
