+++ hr_employee_declaration/static/description/employee_survey_field.png
