+++ hr_employee_declaration/static/description/mail_template_example.png
