+++ hr_employee_declaration/static/description/survey_type.png
