+++ hr_employee_declaration/views/hr_employee_with_declaration_settings.xml
