+++ hr_employee_declaration/views/survey_type_with_is_employee_declaration.xml
